+++ cli/Sources/TuistSupport/Constants.swift
@@ -1,7 +1,7 @@
 import Foundation
 
 public enum Constants {
-    @TaskLocal public static var version: String! = "4.130.3"
+    @TaskLocal public static var version: String! = "4.133.0"
     public static let versionFileName = ".tuist-version"
     public static let binFolderName = ".tuist-bin"
     public static let binName = "tuist"