+++ server/lib/tuist_web/plugs/loader_plug.ex
@@ -10,7 +10,7 @@ defmodule TuistWeb.Plugs.LoaderPlug do
   alias Tuist.Projects
   alias TuistWeb.Errors.BadRequestError
   alias TuistWeb.Errors.NotFoundError
-  alias TuistWeb.Plugs.AppsignalAttributionPlug
+  alias TuistWeb.Plugs.SentryContextPlug
 
   def init([]), do: [:project, :account, :run]
   def init(opts), do: opts
@@ -56,7 +56,7 @@ defmodule TuistWeb.Plugs.LoaderPlug do
       account ->
         conn
         |> assign(:selected_account, account)
-        |> AppsignalAttributionPlug.set_selection_tags()
+        |> SentryContextPlug.set_selection_context()
     end
   end
 
@@ -86,7 +86,7 @@ defmodule TuistWeb.Plugs.LoaderPlug do
         |> assign(:selected_account, project.account)
         |> assign(:selected_project, project)
         |> assign(:selected_run, run)
-        |> AppsignalAttributionPlug.set_selection_tags()
+        |> SentryContextPlug.set_selection_context()
 
       {:error, :not_found} ->
         raise NotFoundError,
@@ -105,7 +105,7 @@ defmodule TuistWeb.Plugs.LoaderPlug do
         conn
         |> assign(:selected_project, project)
         |> assign(:selected_account, project.account)
-        |> AppsignalAttributionPlug.set_selection_tags()
+        |> SentryContextPlug.set_selection_context()
 
       {:error, :not_found} ->
         raise NotFoundError,